ml_runtimes/custom_scikit/requirements.txt
click==6.7
Flask==0.12.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==0.23
numpy>=1.12.1
pandas>=0.20
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
Werkzeug==0.12.2
nltk>=3.2.1
tqdm>=4.11.2
tabulate>=0.7.7
pyahocorasick>=1.1.4
